Change qpack_max_table_capacity and qpack_blocked_streams to use smaller size integer types.
qpack_max_table_capacity has a maximum of 2^30 - 1 which fits inside an
unsigned int. qpack_blocked_streams has a maximum of 2^16 - 1 which fits
inside an unsigned short. Using these types instead of unsigned longs
reduces the room for error (setting the values too high) and reduces the
size of h3c_frame_t since h3c_frame_settings_t was (and still is) the
largest struct in the union of h3c_frame_t.

+++ h3c/include/h3c/frame.h
@@ -55,8 +55,8 @@ typedef struct {
 typedef struct {
   uint64_t max_header_list_size;
   uint64_t num_placeholders;
-  uint64_t qpack_max_table_capacity; // Max: 2^30 - 1
-  uint64_t qpack_blocked_streams; // Max: 2^16 - 1
+  uint32_t qpack_max_table_capacity; // Max: 2^30 - 1
+  uint16_t qpack_blocked_streams; // Max: 2^16 - 1
 } h3c_frame_settings_t;
 
 H3C_EXPORT extern const h3c_frame_settings_t h3c_frame_settings_default;
